Updated config and move submodules
