+++ scripts/search.mjs
@@ -1,6 +1,6 @@
 import { writeFileSync } from "fs";
 
-import config from "../content/config.mjs";
+import config from "../config/config.mjs";
 import clientPromise from "../lib/mddb.mjs";
 
 // const omit = (obj = {}, keys = []) => {
